move base classes into base files
