+++ src/rxn_network/core/composition.py
@@ -5,7 +5,13 @@ from pymatgen.core.composition import Composition as PymatgenComposition
 
 
 class Composition(PymatgenComposition):
-    """Customized Composition class adapted from pymatgen"""
+    """
+    Modified composition class adapted from pymatgen.
+
+    The purpose of this is to modify / extend methods for better performance within the
+    rxn_network package.
+
+    """
 
     def __init__(
         self, *args, strict: bool = False, **kwargs
