+++ src/rxn_network/enumerators/base.py
