+++ src/rxn_network/enumerators/basic.py
@@ -12,9 +12,9 @@ from pymatgen.analysis.phase_diagram import GrandPotentialPhaseDiagram, PhaseDia
 from pymatgen.entries.computed_entries import ComputedEntry
 from tqdm import tqdm
 
-from rxn_network.core.enumerator import Enumerator
 from rxn_network.entries.entry_set import GibbsEntrySet
 from rxn_network.entries.utils import initialize_entry
+from rxn_network.enumerators.base import Enumerator
 from rxn_network.enumerators.utils import group_by_chemsys
 from rxn_network.reactions.computed import ComputedReaction
 from rxn_network.reactions.reaction_set import ReactionSet
