+++ src/rxn_network/reactions/base.py
