+++ src/rxn_network/reactions/basic.py
@@ -15,7 +15,7 @@ from monty.fractions import gcd_float
 from pymatgen.core.composition import Element
 
 from rxn_network.core.composition import Composition
-from rxn_network.core.reaction import Reaction
+from rxn_network.reactions.base import Reaction
 
 TOLERANCE = 1e-6  # Tolerance for determining if a particular component fraction is > 0.
