--- tests/builders/test_retrosynthesis.py
@@ -1,10 +1,0 @@
-"""Tests for retrosynthesis builder(s)"""
-
-from rxn_network.builders.retrosynthesis import SynthesisRecipeBuilder
-
-# TODO: Add tests for SynthesisRecipeBuilder
-
-
-def test_synthesis_recipe_builder():
-    """Test SynthesisRecipeBuilder"""
-    _ = SynthesisRecipeBuilder
